Reduce scope of `UberAuthenticating` protocol
